+++ core/src/proof.rs
@@ -1,9 +1,11 @@
 use avail_rust::kate_recovery::{
-	data::Cell,
+	commons::ArkPublicParams,
+	data::{Cell, SingleCell},
 	matrix::{Dimensions, Position},
+	proof::{verify_v2, Error},
 };
+
 use color_eyre::eyre;
-use dusk_plonk::commitment_scheme::kzg10::PublicParameters;
 use futures::future::join_all;
 use itertools::{Either, Itertools};
 use std::sync::Arc;
@@ -13,17 +15,15 @@ use tracing::debug;
 #[cfg(target_arch = "wasm32")]
 use web_time::Instant;
 
-mod core;
-
 use crate::utils::spawn_in_span;
 
 async fn verify_proof(
-	public_parameters: Arc<PublicParameters>,
+	public_parameters: Arc<ArkPublicParams>,
 	dimensions: Dimensions,
 	commitment: [u8; 48],
-	cell: Cell,
-) -> Result<(Position, bool), core::Error> {
-	core::verify(&public_parameters, dimensions, &commitment, &cell)
+	cell: SingleCell,
+) -> Result<(Position, bool), Error> {
+	verify_v2(&public_parameters, dimensions, &commitment, &cell)
 		.map(|verified| (cell.position, verified))
 }
 
@@ -33,23 +33,25 @@ pub async fn verify(
 	dimensions: Dimensions,
 	cells: &[Cell],
 	commitments: &[[u8; 48]],
-	public_parameters: Arc<PublicParameters>,
+	public_parameters: Arc<ArkPublicParams>,
 ) -> eyre::Result<(Vec<Position>, Vec<Position>)> {
 	if cells.is_empty() {
 		return Ok((Vec::new(), Vec::new()));
-	};
+	}
 
 	let start_time = Instant::now();
 
 	let tasks = cells
 		.iter()
-		.map(|cell| {
-			spawn_in_span(verify_proof(
-				public_parameters.clone(),
-				dimensions,
-				commitments[cell.position.row as usize],
-				cell.clone(),
-			))
+		.filter_map(|cell_type| {
+			SingleCell::try_from(cell_type.clone()).ok().map(|cell| {
+				spawn_in_span(verify_proof(
+					public_parameters.clone(),
+					dimensions,
+					commitments[cell.position.row as usize],
+					cell,
+				))
+			})
 		})
 		.collect::<Vec<_>>();
 
@@ -58,14 +60,20 @@ pub async fn verify(
 		.into_iter()
 		.collect::<Result<_, _>>()?;
 
-	let results: Vec<(Position, bool)> = join_results.into_iter().collect::<Result<_, _>>()?;
+	let results: Vec<(Position, bool)> = join_results
+		.into_iter()
+		.map(|r| r.map_err(|e| eyre::eyre!("{:?}", e)))
+		.collect::<Result<_, _>>()?;
 
 	debug!(block_num, duration = ?start_time.elapsed(), "Proof verification completed");
 
 	Ok(results
 		.into_iter()
-		.partition_map(|(position, is_verified)| match is_verified {
-			true => Either::Left(position),
-			false => Either::Right(position),
+		.partition_map(|(position, is_verified)| {
+			if is_verified {
+				Either::Left(position)
+			} else {
+				Either::Right(position)
+			}
 		}))
 }