+++ client/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.13.0]
+
+- Updated `avail-light-core` to 1.2.9
+
 ## [1.12.13](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.13) - 2025-05-30
 
 - Skiped inserts into the DHT when client is permanently in server mode