+++ fat/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [1.12.8]
 
-- Update `avail-light-core` to 1.2.8
+- Update `avail-light-core` to 1.2.9
 
 ## [1.12.7](https://github.com/availproject/avail-light/releases/tag/avail-light-fat-v1.12.7) - 2025-05-12